# API/docs/swagger.yaml
definitions:
  util.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  title: 接口
paths:
  /admin/articles:
    get:
      parameters:
      - description: pageIndex
        in: path
        name: pageIndex
        required: true
        type: integer
      - description: pageSize
        in: path
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
      summary: 获取文章列表
      tags:
      - admin
    post:
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
      summary: 新增文章
      tags:
      - admin
  /admin/articles/:id:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
      summary: 查看详情
      tags:
      - admin
    post:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
      summary: 修改详情
      tags:
      - admin
swagger: "2.0"